layered caching for session stores
This implements `CachingSessionStore`, which holds two `SessionStore`s.
One acts as a frontend cache, while the other asks as the backend
persistence store.

Anything that implements `SessionStore` may be used as either the cache
or the store.

We also refactor our Moka implementation as it can now be used for
layered caching directly in via `CachingSessionStore`.