ensure sessions are deleted when empty (#22)
When a session is empty it should be removed entirely.

Previously this was not the case as the modified flag cannot be set when
the underlying session data is empty. This makes sense because an empty
session should not be updated but instead deleted.

However, this also meant that removing the last key of a previously
stored session would not reflect this update in the store--the session
would not be saved to the store or deleted.

To address this, we now look for an empty session after resolving the
request. When found, we delete the session from the store and return the
response.

Fixes #21.