negative caching is obviated by removal cookies
Empty sessions are deleted from stores which means it's not currently
possible to support negative caching as originally thought. However,
such a deletion also issues a removal cookie, which in turn means
deleted sessions will not be requested from behaving clients. In other
words, the need for negative caching should be largely obviated by this
change.