track loaded sessions to enable concurrent access
This moves to a pattern of tracking loaded sessions via a hash set. By
doing so, we can share sessions concurrently. In other words, we no
longer need a lock to protect the session for the duration of the
request and instead are able to share sessions between concurrent tasks.

Why do we need this at all? A data race can otherwise occur when a
session has been loaded but its modifications have not been stored. For
instance, another task may load the session from the store in the
meantime and this load will overwrite pending modifications as it will
effectively replace the session in memory.

One solution to this problem is to lock the request, as we've done in a
stopgap capacity. This unscalable at present and even when refined, i.e.
to lock per session as opposed for all sessions, limits concurrency over
the session itself. In practice, sessions can only be operated on
serially and are not available concurrently.

A more flexible approach is implemented here, allowing for concurrent
memory access by limiting loads from the store to situations where no
session exists in the loaded set. Put another way, sessions are tracked
such that they can be used instead of loading from the store thus
ensuring that pending modifications are not lost by a load.

See #36 for more details about the stopgap locking behavior; this
replaces that approach altogether.