Fix the mysql url
Diesel uses libmysqlclient internally, which tries to be clever by using
a unix socket instead of a network socket if it
sees a database url that contains `localhost` as a host name. That won't
work in this case, as the db is running in a docker container. This
change switches to use `127.0.0.1` as host instead which forces the
usage of a network socket.