provide session expiration a la django
Here we rework how session expiration works, implementing a new type,
`SessionExpiry` (later to be renamed `Expiry`). This enum defines three
possible types of expiry,

1. Activity durations, i.e. durations for which the session will be
   renewed until inactivity for that duration has elapsed,
2. Absolute expirations, i.e. timestamps at which the session will
   expire regardless of activity,
3. And browser sessions, which are sessions that are valid for so long
   as the client remains open or the default threshold (two weeks) is
   reached.

Several breaking changes are required to make this change and
importantly, the session itself is now serialized to the session store.
This also means we no longer need `SessionRecord` and it has been
removed entirely.

Please note that fields and methods related to "expiration" have been
renamed or removed and with this change all use cases for the expiration
are provided for by the expiry type.

Closes #57.