push active filtering down to stores
Following along with the Django approach, we push the responsibility of
defining an active session semantic down to the stores themselves. Most
stores already filter results based on e.g. the expiry date, but with
this change we allow that semantic to be defined however implementors
might like.

Provided stores that did not already do their own filtering have been
updated to account for this change.