consolidate service configuration
Previously a separate `CookieConfig` was used. However, this
configuration scope extends beyond the cookie and includes e.g. the
session expiry. In fact, the proper scope for this configuration is
either the session or the service. The session does not need to be
coupled to how it might be represented, i.e. as a cookie, so here we
move those concerns onto the service. The config is still contained
within its own type, to make it easier to operate over within the
middleware.