only delete from stores when we have a cookie
If we don't have a cookie, the session is new and therefore has never
been saved. Given this, we should not invoke the overhead of talking to
the store. Likewise, we don't need to set a removal cookie as we have
never set the cookie.

This should address the original issue in #89.