remove support for concurrent access
Here we remove support for accessing the session concurrently.
Previously we supported this via shared memory. However, this approach
only scales via sticky session proxying, or similiar approaches and
requires the overhead of e.g. `dashmap`.

In the future, we may be able to explore enabling this by allowing the
data of the session itself be a generic type, constrained by a trait.
Doing so could allow us to configure distributed shared memory.