implement lazy sessions
Here we've reworked our session implementation to avoid loading sessions
until they're actually used. In other words, these are "lazy" sessions.

We enable this by colocating the session store with the session.

Doing so provides some noteworthy benefits:

1. The overhead of loading from a backend store in the middleware is
   entirely eliminated--this crate is "infinitely" faster now insomuch
   as the middleware itself is concerned.
2. Because stores are colocated with sessions, we are now able to read
   from and write to sessions from e.g. handlers--this will allow us to
   address concurrent data access with some backends, but also relates
   to the first point: we only pay for what we actually use.

However, this also introduces some important changes to the interface,
most obviously the fact that session methods are now async and fallible
(for stores that can result in an error).

Ultimately moving in this direction ensures we can scale the middleware.
In other words, this is likely needed to claim "web scale" (for some
measure of that dubious term). Prior to this we've worked around the
load cost with e.g. caching stores, but at some point it becomes
desirable to simply not load unless the session is used.

We must also point out that this is how Django has designed their
session middleware[1]: at no point is the session loaded and instead this
is cleverly deferred by leaning on Python's magic methods and property
setters.

As per ergonomics, because async-await is ubiquitious in our API, there
is a general degradation to note here: we must now await all these
methods. At the same time, the proliferation of await points should be
beneficial to highly concurrent applications.

That said, this design is generally easier to use correctly because the
costs are moved to the public API callsites as opposed to the prior
design, which obscured specifically the load cost within the middleware.

Finally it's also important to point out that session stores are now
used as trait objects, which means that the error type has been changed:
it's now a concrete type owned by this crate. Implementers should map
their backend errors to this new error enum.

For more details on the exploration of this design, please see this
discussion thread: https://github.com/maxcountryman/tower-sessions/discussions/101

[1]: https://github.com/django/django/blob/0ee2b8c326d47387bacb713a3ab369fa9a7a22ee/django/contrib/sessions/middleware.py