make service infallible
This reworks the tower service such that we always return a response. In
other words, the service becomes infallible. What this means is we no
longer need to build an intermediary service that captures boxed errors
when using this crate with axum.

One side effect of this is that we lose the ability to handle specific
errors, such as the session save failing, via downcasting. Instead, the
middleware will return a 500 response. To address this, we could extend
this implementation to allow for an on error callback, which could take
an error and return a response.

This is a breaking change which also removes two extraneous variants
from the session error type:

- `MissingId`
- `MissingCookies`