Improve session ID (#141)
* Exchange UUID for nanoid

UUIDv4 has 122 random bits, with 6 fixed bits for identifying the
format. Additionally, it uses hexadecimal encoding, which makes it take
more bandwidth to transfer.

This patch replaces UUIDv4 with an array of bytes derived from the
nanoid crate, which gives 66 bits of entropy. This exceeds the minimum
recommended by OWASP. Additionally, the byte array is encoded as
URL-safe Base64 for usage as a string.

* Switch to Rand/Base64 crates

Improves memory use slightly, as well as type safety.

* Correct documentation typo