session cookie signing and encryption
This provides session cookie signing and encryption via the feature flags, `signed` and `private` respectively.

Cryptographic treatment of the cookie value provides some additional properties that may be beneficial to an application's security posture. For example, while session keys are already composed of crytographically secure random values, and therefore exceedingly difficult for an attacker to guess, the addition of a cryptographic key presents another dimension an attacker must overcome in order to successfully guess a valid session.

Furthermore, while session stores may themselves be secured, e.g. via techniques like encryption at rest or session store implementors may provide measures to ensure plaintext values are not stored directly, signed or encrypted cookies again require knowing both the contents of a store and the cryptographic key (which is hopefully not colocated in the same store).

Altogether, signing and encryption are "defense in depth" techniques which can increase the security posture of an application.

Finally such cryptographic treatment is not necessary in all contexts and so is enabled here via an opt-in feature. When the flags are enabled, users must invoke the `with_signed` or `with_private` methods to ensure signing or encryption are leveraged.